Implement pipeline module
We can now define an operaiton by subclassing one of the
- Task
- TrainerTask
- HyperParameterTuningTask

Notice that the operaiton class should be named in CamelCase and suffix
with `Op`, also its name should corresponding to the python script it
run. For example TrainModelOp will run train-model.py

To use the operation class, we just call it with the keyword arguments.